fix delete user with providers

+++ api/helpers/_identityaccessmanager.py
@@ -25,6 +25,7 @@ from api.utils.configuration import configuration
 from api.utils.exceptions import (
     DeleteOrganizationWithUsersException,
     DeleteRoleWithUsersException,
+    DeleteUserWithProvidersException,
     DeleteUserWithRoutersException,
     InvalidCurrentPasswordException,
     InvalidTokenExpirationException,
@@ -277,8 +278,11 @@ class IdentityAccessManager:
 
         try:
             await postgres_session.execute(statement=delete(table=UserTable).where(UserTable.id == user_id))
-        except IntegrityError:
-            raise DeleteUserWithRoutersException()
+        except IntegrityError as e:
+            if "provider_user_id_fkey" in str(e.orig):
+                raise DeleteUserWithProvidersException()
+            if "router_user_id_fkey" in str(e.orig):
+                raise DeleteUserWithRoutersException()
         await postgres_session.commit()
 
     async def update_user(
